Fix for FL-3677 and Fl-3679"
Gentoo decided to rename:
dev-perl/gtk2-perl -> dev-perl/Gtk2

consequently, all the sharded ebuilds are
now broken.

To fix this, I've created a "virtual" ebuild
that pulls in dev-perl/Gtk2 to address the
issue.

It's a good fix because as the fixes hit the
shards at different intervals, the ebuilds
will transparently switch to the new Gtk2
dependency. Once all the ebuilds in the portage
tree are moved to the new dependency, this virtual
can be removed.